Add tsup for worker bundling and update package dependencies
